test(parsing): split merchant provider specs
